%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Device.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{bringToTop.Device}
\alias{bringToTop.Device}
\alias{Device.bringToTop}
\alias{bringToTop.Device}
\alias{bringToTop,Device-method}


 \title{Bring a device to top}

 \description{
  Bring a device to top. By default it is the current active device that will
  be brought to top.
 }

 \synopsis{bringToTop.Device(static, which=dev.cur(), ...)}
\usage{Device$bringToTop(which=dev.cur(), ...)}

 \arguments{
   \item{which}{Index (an \code{\link[base]{integer}}) of device.}
   \item{...}{Not used.}
 }

 \value{
  Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Device}}.
 }
\keyword{internal}
\keyword{methods}
