%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Device.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{choose.Device}
\alias{choose.Device}
\alias{Device.choose}
\alias{choose.Device}
\alias{choose,Device-method}


 \title{GUI figure selector}

 \description{
  Opens a window with a list of open figures. By clicking on any figure
  in the list, that figure is activated and brought to front.
 }

 \usage{\method{choose}{Device}(static, ...)}

 \arguments{
   \item{...}{Not used.}
 }

 \value{Returns nothing.}

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Device}}.
 }
\keyword{internal}
\keyword{methods}
