%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Device.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getPrintPath.Device}
\alias{getPrintPath.Device}
\alias{Device.getPrintPath}
\alias{getPrintPath.Device}
\alias{getPrintPath,Device-method}


 \title{Gets the current path where figures are saved}

 \description{
  Gets the current path where figures are saved.
 }

 \usage{\method{getPrintPath}{Device}(this, ...)}

 \arguments{
   \item{...}{Not used.}
 }

 \value{Returns nothing.}

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.graphics:setPrintPath.Device]{*setPrintPath}()}.
 }

 \keyword{dplot}
\keyword{internal}
\keyword{methods}
