%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Device.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{isOpen.Device}
\alias{isOpen.Device}
\alias{Device.isOpen}
\alias{isOpen.Device}
\alias{isOpen,Device-method}


 \title{Checks if a device is open or not}

 \description{
  Checks if a device is open or not.
 }

 \synopsis{isOpen.Device(con, ...)}
\usage{Device$isOpen(...)}

 \arguments{
   \item{...}{Use 'index' as the index of the device.}
 }

 \value{
   Returns \code{\link[base:logical]{TRUE}} if the device is open, otherwise \code{\link[base:logical]{FALSE}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[grDevices:dev]{dev.list}}.

   For more information see \code{\link{Device}}.
 }
\keyword{internal}
\keyword{methods}
