%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Colors
\alias{colorsToRGB}
\alias{getBlue}
\alias{getBlue.Colors}
\alias{getColorNames}
\alias{getCyanMagentaColors}
\alias{getGray}
\alias{getGreen}
\alias{getGreen.Colors}
\alias{getHeatColors}
\alias{getHSV}
\alias{getMonochrome}
\alias{getMonochrome.Colors}
\alias{getRainbowColors}
\alias{getRed}
\alias{getRed.Colors}
\alias{getRGB}
\alias{getRgbFromWavelength}
\alias{getRGBorHSV}
\alias{getRGBorHSV.Colors}
\alias{getTerrainColors}
\alias{getTopoColors}
\alias{invert}
\alias{invert.Colors}
\alias{isColorName}
\alias{rescale1D}
\alias{rescale1D.Colors}
\alias{rgbToColors}
\alias{toGrayScale}
\alias{toGrayScale.Colors}
\alias{toMonochrome}
\alias{toMonochrome.Colors}

% Device
\alias{bringToTop}
\alias{bringToTop.default}
\alias{callFunction}
\alias{callFunction.Device}
\alias{choose}
\alias{choose.default}
\alias{closeAll}
\alias{constructor.Device}
\alias{finalize.Device}
\alias{getAspectRatio}
\alias{getAspectRatio.Device}
\alias{getDimensions}
\alias{getDimensions.Device}
\alias{getMouseClick}
\alias{getMouseClick.Device}
\alias{getNext}
\alias{getOptions}
\alias{getParameters}
\alias{getPlotParameters}
\alias{getPlotParameters.Device}
\alias{getPrintPath}
\alias{isOpen}
\alias{isOpen.default}
\alias{new}
\alias{new.default}
\alias{paste}
\alias{paste.default}
\alias{plotRegions}
\alias{putAuthor}
\alias{putAuthor.Device}
\alias{putDataset}
\alias{putDataset.Device}
\alias{putTimestamp}
\alias{putTimestamp.Device}
\alias{resetOptions}
\alias{set}
\alias{setOptions}
\alias{setPlotParameters}
\alias{setPlotParameters.Device}
\alias{setPrintPath}

% GifAnimator
\alias{GifAnimator}
\alias{as.character.GifAnimator}
\alias{clearRecord}
\alias{clearRecord.GifAnimator}
\alias{combineFrames}
\alias{combineFrames.GifAnimator}
\alias{finalize.GifAnimator}
\alias{findPngDevice}
\alias{getCurrentFrame}
\alias{getCurrentFrame.GifAnimator}
\alias{getCurrentFrameName}
\alias{getCurrentFrameName.GifAnimator}
\alias{getFrames}
\alias{getFrames.GifAnimator}
\alias{getPathname}
\alias{getPathname.GifAnimator}
\alias{getRecordFormat}
\alias{getRecordFormat.GifAnimator}
\alias{getRecordPath}
\alias{getRecordPath.GifAnimator}
\alias{isFrameRecorded}
\alias{isFrameRecorded.GifAnimator}
\alias{recordClose}
\alias{recordClose.GifAnimator}
\alias{recordCurrent}
\alias{recordCurrent.GifAnimator}
\alias{recordOpen}
\alias{recordOpen.GifAnimator}
\alias{reset}
\alias{reset.GifAnimator}
\alias{seq.GifAnimator}
\alias{setRecordPath}
\alias{setRecordPath.GifAnimator}
\alias{skip}
\alias{skip.GifAnimator}

% Plots
\alias{Plots}
\alias{getAbundantClosePoints}
\alias{getAbundantClosePoints.Plots}
\alias{getClosePoints}
\alias{plotSymbols}
\alias{plotSymbols.Plots}
\alias{setStyle}
\alias{setStyle.Plots}
\alias{subplots}
\alias{subplots.default}
\alias{subplots.Plots}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
