%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Device.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{open.Device}
\alias{open.Device}
\alias{Device.open}
\alias{open.Device}
\alias{open,Device-method}


 \title{Opens a device}

 \description{
  Static method to open a device, where the type is infered from the
  filename.
 }

 \synopsis{open.Device(con, ...)}
\usage{Device$open(...)}

 \arguments{
   \item{...}{Arguments passed to \code{\link[R.graphics:getParameters.Device]{*getParameters}()}.}
 }

 \value{Returns nothing.}

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.graphics:getParameters.Device]{*getParameters}()}.
   \code{\link[R.graphics:print.Device]{*print}()}.

   For more information see \code{\link{Device}}.
 }
\keyword{internal}
\keyword{methods}
