%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Device.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setPrintPath.Device}
\alias{setPrintPath.Device}
\alias{Device.setPrintPath}
\alias{setPrintPath.Device}
\alias{setPrintPath,Device-method}


 \title{Sets the current path where figures are to be saved}

 \description{
  Sets the current path where figures are to be saved.
 }

 \usage{\method{setPrintPath}{Device}(this, path, ...)}

 \arguments{
   \item{path}{Filename path where figures are saved.}
   \item{...}{Not used.}
 }

 \value{Returns nothing.}

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.graphics:setPrintPath.Device]{*setPrintPath}()}.
 }

 \keyword{dplot}
\keyword{internal}
\keyword{methods}
